Fixed MSAA bug
If one were to force one sample for MSAA, an error is triggered. The
code was changed to account for one sample by not implementing the extra
framebuffer attachment (the no MSAA case).

In addition a minor change was made in some of the shader code. By
convention when dealing with more advanced illumination techniques,
convention dictates that light rays always point away from the object
being illuminated. Even though the code simply involved a dot product
with a single directional source for a lambertian model, this convention
was adapted for the sake of consistency with future potential samples.